+++ backend/src/main/java/com/techpro/finalproject/exception/ItemNotFoundException.java
@@ -2,7 +2,7 @@ package com.techpro.finalproject.exception;
 
 public class ItemNotFoundException extends RuntimeException{
 
-    public ItemNotFoundException(Long id){
+    public ItemNotFoundException(Integer id){
         super("Could not find the item with id: "+id);
     }
 }
