+++ backend/src/main/java/com/techpro/finalproject/exception/OrderDetailsNotFoundException.java
@@ -1,7 +1,7 @@
 package com.techpro.finalproject.exception;
 
 public class OrderDetailsNotFoundException extends RuntimeException{
-    public OrderDetailsNotFoundException(Long id){
+    public OrderDetailsNotFoundException(Integer id){
         super("Could not find the order details with id: "+id);
     }
 }
