+++ backend/src/main/java/com/techpro/finalproject/exception/OrderNotFoundException.java
@@ -1,7 +1,7 @@
 package com.techpro.finalproject.exception;
 
 public class OrderNotFoundException extends RuntimeException{
-    public OrderNotFoundException(Long id){
+    public OrderNotFoundException(Integer id){
         super("Could not find the order with id: "+id);
     }
 }
