+++ backend/src/main/java/com/techpro/finalproject/exception/PersonNotFoundException.java
@@ -1,7 +1,7 @@
 package com.techpro.finalproject.exception;
 
 public class PersonNotFoundException extends RuntimeException{
-    public PersonNotFoundException(Long id){
+    public PersonNotFoundException(Integer id){
         super("Could not find the person with id: "+id);
     }
 }
