+++ backend/src/main/java/com/techpro/finalproject/repository/ItemRepository.java
@@ -3,5 +3,5 @@ package com.techpro.finalproject.repository;
 import com.techpro.finalproject.model.Item;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface ItemRepository extends JpaRepository<Item, Long> {
+public interface ItemRepository extends JpaRepository<Item, Integer> {
 }
