+++ backend/src/main/java/com/techpro/finalproject/repository/OrderDetailsRepository.java
@@ -1,7 +1,9 @@
 package com.techpro.finalproject.repository;
 
 import com.techpro.finalproject.model.OrderDetails;
+
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface OrderDetailsRepository extends JpaRepository<OrderDetails, Long> {
+public interface OrderDetailsRepository extends JpaRepository<OrderDetails, Integer> {
+
 }
