+++ backend/src/main/java/com/techpro/finalproject/repository/PeopleRepository.java
@@ -3,5 +3,5 @@ package com.techpro.finalproject.repository;
 import com.techpro.finalproject.model.People;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface PeopleRepository extends JpaRepository<People,Long> {
+public interface PeopleRepository extends JpaRepository<People,Integer> {
 }
